WIP commit with ftss and semantic
