+++ src/bikidata/__init__.py
@@ -1,2 +1,9 @@
-from .main import query, build, build_ftss
+from .main import (
+    build,
+    build_ftss,
+    build_from_iterator,
+    decode_unicode_escapes,
+    literal_to_parts,
+)
 from .semantic import build_semantic
+from .query import query, raw
